Moved mem utilities to their own file
