+++ CDargonQuest/CDargonQuest.vcxproj.filters
@@ -153,6 +153,9 @@
     <ClCompile Include="log.c">
       <Filter>Source Files\Utilities</Filter>
     </ClCompile>
+    <ClCompile Include="mem_util.c">
+      <Filter>Source Files\Utilities</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="common.h">
@@ -291,6 +294,9 @@
       <Filter>Header Files\Utilities</Filter>
     </ClInclude>
     <ClInclude Include="math_util.h">
+      <Filter>Header Files\Utilities</Filter>
+    </ClInclude>
+    <ClInclude Include="mem_util.h">
       <Filter>Header Files\Utilities</Filter>
     </ClInclude>
   </ItemGroup>
